Refactor password reset functionality: move to AuthController and update security configurations
