+++ src/main/java/lk/sliit/itpm/demo/controller/AdminController.java
@@ -51,13 +51,5 @@ public class AdminController {
         return userService.updateUser(id, user);
     }
 
-    @PostMapping("/reset-password")  // Remove leading slash
-    public ResponseEntity<Void> resetPassword(@RequestBody JsonNode jsonNode) {
-        String email = jsonNode.get("email").asText();
-        String newPassword = jsonNode.get("newPassword").asText();
-
-        userService.resetPassword(email, newPassword);
-        return ResponseEntity.ok().build();
-    }
 
 }
