+++ src/main/java/lk/sliit/itpm/demo/controller/AuthController.java
@@ -1,0 +1,27 @@
+package lk.sliit.itpm.demo.controller;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import lk.sliit.itpm.demo.service.UserService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/auth")
+@CrossOrigin(origins = "*", allowedHeaders = "*")
+public class AuthController {
+
+    private final UserService userService;
+
+    public AuthController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @PostMapping("/reset-password")
+    public ResponseEntity<Void> resetPassword(@RequestBody JsonNode jsonNode) {
+        String email = jsonNode.get("email").asText();
+        String newPassword = jsonNode.get("newPassword").asText();
+        
+        userService.resetPassword(email, newPassword);
+        return ResponseEntity.ok().build();
+    }
+}
